+++ src/components/sidebar.tsx
@@ -185,7 +185,7 @@ export default function SideBar() {
             </div>
           )}
           <p className="ml5">
-            Metahkg build {process.env.REACT_APP_build || "v0.5.6rc1"}
+            Metahkg build {process.env.REACT_APP_build || "v0.5.6rc2"}
           </p>
         </Box>
       </Drawer>